+++ layout/_partial/main/float_tools.njk
@@ -5,12 +5,7 @@
   </a>
 
   {# 搜索 #}
-  {% if theme.search.enable and theme.search.type === 'algolia' %}
-    <a class="sea-float-tools-btn" id="sea-search-icon">
-      <i class="sea-font">search</i>
-    </a>
-    <div id="sea-search-input"></div>
-  {% endif %}
+  {{ generateSearchHtml() | safe }}
 
   {# 主题切换 #}
   <div class="sea-float-tools-btn" id="sea-theme-mode">
